requirements.txt
asyncpg
fastapi
uvicorn[standard]
pydantic-settings
python-multipart
httpx
pytest
psycopg2